requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.6
phonenumbers==8.13.33
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydub==0.25.1
PyJWT==2.8.0
requests==2.31.0
SpeechRecognition==3.10.3
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
